feat: add formation video player and layout
Implement FormationLayout and Formation components to provide a video 
player experience. The FormationLayout sets up the navigation stack, 
while the Formation component the expo-video library to 
play a video with controls for play pause. Update app.json tosupport background playback and picture-in-picture features. Add 
Formation tab to the main layout for easy access.